requirements.txt
pandas
numpy
rapidfuzz
ipykernel
xgboost
scikit-learn==1.5.2
matplotlib
jupyter
GitPython
psycopg2-binary==2.9.10
